1.5.1

+++ Bot/Modules/WeatherReportCurrent.cs
@@ -45,7 +45,7 @@ namespace Bot
                 public double Temp { get; set; }
 
                 [JsonProperty("pressure")]
-                public int Pressure { get; set; }
+                public float Pressure { get; set; }
 
                 [JsonProperty("humidity")]
                 public int Humidity { get; set; }
